Add skip_preprocess flag and usage sample
This introduces the way for skipping preprocess sequence to
pass pure string for encode into the backend engine, and also shows
a sample code for format conversion from pure encoded data when using
non-preprocess sequence.

Why do we need this:
Sometimes a backend engine want to recieve a natural string (i.e. char*)
for an encode request. However, the engine have to merge the string splited
by preprocess in current implementation. This behavior may cause redundant
memory copy to merge the string. Or it may force to rewrite the engine's
interface. To prevent this, this patch introduces the way to skip the preprocess.
This patch includes also an usage sample code called "null_naive" taking the
argument by char*. It works really like "null" backend engine.