Fix reconstruct failure with lack of chksum type
This patch fixes reconstruct to set correct checksum type.

Changes of the checksum type definition (on Commit 6d9c09fa)
results in a reconstuct test failure because current
liberasurecode_reconstruct_fragment() will fill the chksum_type
with 0 in the fragment header. However, we need to set the
chksum_type correctly (e.g. 1 at CHKSUM_NONE as a default) on
reconstruct as same as encode.

This patch makes liberasurecode to ensure to set chksum_type
into the fragment header on reconstruct and fixes the bug.